page: Run HTMLCacheUpdateJob on delete for pages transclude the page
When a template is deleted there is no HTMLCacheUpdateJob to invalidate
the cache of each page transcluding the deleted page. This avoids that
the deleted content gets invisible in a short time after deletion.
Only a RefreshLinksJob is running to reparse each page, but that can
take more time as the invalidation, so use the invalidation first.
This is similiar to page creation.

Bug: T275747
Change-Id: I95ae7476b24ae41ef3c805411bd34d234f8aa5b6