Deprecate Title::getBrokenLinksFrom()
This is only used SpecialBrokenRedirects where it doesn't need it. If a
redirect is not in the redirect table, we shouldn't look for it in
pagelinks table. This is also adding complexity for migrating pagelinks
to the new schema.

Tested locally and confirm it doesn't affect anything to my knowledge.

Bug: T346457
Change-Id: I4e570c29da7e0b087f074385b0c132397d1ca540