Fix leakage of ipblocks rows from BlockUserTest
This fixes a test failure in a subsequent commit.

Also delete a test case from testReblock() which fails as a result. It
seems to just be running the exact same test twice, and expecting it to
fail the first time and succeed the second time. Unclear why it would do
that.

Change-Id: I3521792a9bed655906dad831f8889f6e0dad23cc