config: Use stale cache instead of hard fail on "no retry" fetch error
Inspired by investigation for T349376 and T346971, although not
currently known to happen in production.

If the backend fetch fails with an error and an HTTP status code that
tells us not to re-try again in the same process, this shouldn't
affect whether or not we use the stale data.

Due to this line doing an early return for non-retryable errors,
it was bypassing the code below this where is_array results in stale
cache being used. Note that $data is only re-assigned from the cache
hit in case of a non-error etcd response.

The added test fails (expected: from-cache-expired, actual: error)
if the EtcdConfig change is not applied.

Change-Id: I9b85c206d929706d692bb46efe034be943156031