SVGReader: Detect CSS animated SVGs
This does a string check on animated and @keyframe in style elements
for SVGs. If there is a match it will treat this SVG as animated.

While in theory this could be a false positive, without fulling
parsing the CSS, the chances of that seem minor and there are no
real negative consequences either.

Bug: T332790
Change-Id: I4b8b0781e8f9135d9ab856f3ec06f5a76c66c9a8