Remove angle brackets from specialmute-email-footer
The angle brackets were probably added to make sure
that the URL doesn't get mixed with other characters.
However, the CI suspects that the angle brackets are HTML
and requires a manual review of every translation update
that includes the translation of this message, such as
I5780dcfd4d958e6c239a9b0448bece511899f03e
(a translation of this message was added in vep.json).

This change moves the URL to a separate line, which
is supposed to be enough for isolation.

Bug: T355306
Change-Id: Ied4007a9c665a66fc7b523fee8ebb2256fdaf69d