[poolcounter] Remove incomplete workaround for T48014
In relatively-ancient times, there was an issue with stale content being
added to the parser cache when reverts were done very quickly after new
content was added.  This seemed to be a race condition between two slow
parses, and code was added to set the cache time "as soon as possible"
so that the render times more closely matched the revision order.
However, this still doesn't *guarantee* that render times matched the
revision order, and eventually a subsequent patch was added to add
the revision ID to the parser cache, which apparently fixed the root
cause.

I72c5e6f86b7f081ab5ce7a56f5365d2f75067a78 sets the render time early
in the ContentRenderer, which does effectively the same thing this
patch did, so the workaround added here is no longer necessary.

Bug: T48014
Bug: T350538
Change-Id: Ide67527158d0a1e517bbc2c693aae15ca80f5618