LinksUpdate: Compare raw domain and path for externallinks table
Using LinkFilter functions can repair some values, but that would never
removes invalid values from the externallinks table.

When LinkFilter get bug fixes on how to handle domain or path there is
no way to refresh the links on the externallinks table (even not with
refreshlinks.php) as it compares the url rebuilded from the database
with the value from the parser. But rebuilding the url from database
must not match, when reversing the url to find back the database row.
Just use the raw values internally.

Bug: T350476
Change-Id: I60e61c229c6923427585a445a08e6130d265519d