Plain text content model should show line breaks
Before this patch, if you go to Special:ChangeContentModel and
set the content model of a page to plaintext, and it has line
breaks, and you go view the page, you cannot see the line breaks.
All the text is pushed together.

This patch fixes that by wrapping everything in <pre>. nl2br was
also considered, but wrapping in <pre> transforms the HTML less,
so is a simpler solution.

Bug: T202424
Change-Id: Iaa08c24f19267e48a92083e5dc4b886d0fe80288