SpecialLonelyPages: Add support for pagelinks read new
First, assume that templatelinks migration is over and simplify the
query. Then you can copy-paste the templatelinks conditions back to the
main query but thankfully since due to templatelinks migration we join
with linktarget, it gets simplified a lot (and much faster in production
once turned on).

Tested both read old and read new locally and it works fine.

Bug: T346457
Change-Id: I15deba9b4096be1b01018562e65146952bab5843