Fix editRecovery data not being deleted on save
In the current code we check if section was not passed
to the function and the key is a empty string. However
the to check if the section is not passed in, we are
checking if section === null, this is incorrect since
if the section is not passed in, it will be undefined
and Javascript thinks undefined !== null (🤯) making this
peice of code break and never delete data if no section
is provided.

In this fix, the check is changed to check if the section
provided is falsy which should cover this specific edge case.

Bug: T351090
Change-Id: I0cf1ceecde3f0ff8d40bed5aaf19dbbf0f2ae1bc