Make isWellFormedLanguageTag() actually case insensitive
BCP 47 language tags are case insensitive. At the moment,
isWellFormedLanguageTag() lowercases the language tag to be checked, but
the regex contains uppercase characters, which means it will incorrectly
return false for a language tag like "en-GB-oed".

This changes the regex to use lowercase letters throughout, for
consistency, since it doesn't use capital letters in other places where
they would normally be used, but also makes it do a case-insensitive
match instead of trying to make sure the case for the regex and language
tag are the same. This should reduce the chance of it breaking if
someone later re-adds capital letters to the regex.

Change-Id: Iacf87d37c7ac515c350399f99e05e3f5bace9b90