Restrict autoloading of mediawiki ui button styles
This was deprecated in 1.26.
We should stop loading this on the article namespace as it means
any article can force the loading of the large mediawiki ui button
to render a single button. It can also be used by vandals to render
fake login buttons.

This only impacts 81 pages (when you exclude subpages) so the fallout
is very manageable and allows us to experiment with alternatives as part of
the discussion in T346469.

Bug: T182050
Change-Id: Ie4aa65ffcf99e53d2786bfe80ea5a18b2a1f3fd5