stats: add metric stats.statslib.buffered
Records the number of samples of all metrics in the buffered statslib
cache. Necessitates adding getSampleCount() to BaseMetric and wiring
access through each metric type.

Moved StatsFactory->flush() to execute before
emitBufferedStatsdData().  This is so the metric generated by
copyToStatsdAt() is sent with the rest of the statsd metrics.

Updated tests.

This metric will be used to track the migration progress towards
statslib.

Bug: T350638
Change-Id: I070f3e30a9b8cd0b0ab44f345207e6f25e602fbb