tests: Remove dead code from WikiPageDbTest::assertPreparedEditNotEquals
PreparedEdit->timestamp is not declared.
There is __get() support to read the property, but not __set() to set
the property.
Setting the timestamp on the clone results in creation of a dynamic
property, that is deprecated in php8.2 and make the test fail.
Remove the dead code, the assertion of not-equal also not test this
property as there is no PreparedEdit::equals() function.

Change-Id: Ic6218b6357a18eedd95b6bfc73b178b5f784447e