Guard against deleted files in MWDebug::getFilesIncluded
When an included file has since been deleted, filesize will raise a PHP
warning and return false, thus causing logspam (possibly) and test
failures (surely).

Add a check for deleted files and report their size as 0 (in case
anything expects the value to be numeric), but keep including them in
the generated debug information in case this turns out to be useful.

Bug: T351986
Change-Id: Ie91c9f2882f12fd8ea822529528dd1dd9efe89db