editpage: Minimize code dealing with TextConflictHelpers
The idea is to minimize the interface between core and the
TwoColConflict extension a little.

This also fixes an issue where a second call would destroy what the
first call did, most certainly resulting in an instable state where
two conflicting extensions are fighting each other.

Luckily this is only a hypothetical scenario as there is only one
known caller:
https://codesearch.wmcloud.org/search/?q=setEditConflictHelperFactory

Change-Id: I6a3f99682f084965cab0e497f63879ac1d091c43