EditPage: Remove unnecessary extra check for blocks
An extra check was added in 7ca9c71bb91128ad9517e43532636a7ba4b1fb7e,
in case permissions were being checked for a user other than the
global context user, because in that case blocks are only checked for
that user's account, and not for the request (including IP, XFF,
cookies). (One use-case was when editing for the first time, just
before a temporary account is created, in which case the permissions
checks are performed against a placeholder temporary user).

No longer needed since 9e7c9c13b900fb822af6a748d9d59f32d686ca52,
after which the PermissionManager checks blocks for the request,
regardless of the user.

Bug: T350116
Change-Id: I8e0478e3c4d10eaa71d48efe69de4d7832aab81d