Edit Recovery: Store better original form data
In order to avoid saving the browser's auto-saved data as the
'original' data of the form, store instead the
defaultValue or defaultChecked of all form fields.

Also use HHTMLFormElement.reset() to discard the Edit Recovery
data rather than loading in the stored original data (even
though these should now be the same, using reset() is clearer).

Bug: T351821
Change-Id: I1ec757d5ddd9f0db66496c6aaef70747d93a5c83