Prioritize search index over page namespace index
I was getting terrible performance on searches, and discovered that the
query planner was first using the page namespace index to quickly find
pages with the right namespaces, then doing a separate full text search
for each page. To fix this durably, use CROSS JOIN instead of "," for
the join, which is semantically the same but acts as a hint to the query
planner that indexes on the first table listed should be searched first.
https://www.sqlite.org/optoverview.html#manual_control_of_query_plans_using_cross_join

Change-Id: I756f6b43f84894c76e46b3a3221107d9834a493a