rdbms: make Query only track tables that are the target of writes
Rename Query::getTables and related variables to reflect this.

This avoids excess complexity and poorly implemented concepts.

Previously, MediaWikiIntegrationTestCaseSchema only passed for mysql
due to duplicateTableStructure() giving the reference table name as
a "changed" table and that happening to be the unprefixed table name
for $wgDBprefix = ''. Amusingly, this worked around the double-prefixing
in getTempTableWrites() since one of the tables was just the unprefixed
version, so it matched up with the sessionTempTables from the call to
duplicateTableStructure(). Since duplicateTableStructure() only takes
unquoted table names with '.' qualifiers, prefix the table name with a
dot when giving it to Query(). This makes tableName() strip the dot
but avoid adding a prefix (note that mysql would not care even with
the dot though postgres would error out).

Bug: T350330
Change-Id: I82893477ad378b99b5aa75b604ab771b2b3cfee7