Remove dataType from mw.Rest AJAX requests
mw.Rest currently locks in the data type of a request to JSON because
`dataType` is set to "JSON" for POST, PUT, and DELETE requests. This
causes a response with Content-Type `text/html` to be incorrectly
converted into JSON, causing the request to reject with a SyntaxError
on runtime.

Removing `dataType` allows JQuery to automatically guess the correct
data to return from the function using the `Content-Type` header
received from the request. This requires all REST requests from
MediaWiki to return accurate `Content-Type` headers, but also allows
non-JSON data to be returned through the REST API.

Bug: T352546
Change-Id: I381f33dd40bd11d89e8843bf412531216be7c725