Add Content-Type header checks to REST API tests
Add extra checks on all REST API tests which return sucess data values.
This extra check is used to ensure that an accurate `Content-Type` is
being sent to the client, as an incorrect header can cause consumers
to misinterpret the data returned by the API (particularly JQuery, as
described in T352546).

The check was added only to endpoints which return success data.
Tests that check for 4xx status codes do not have the content type
check.

This is a follow-up to I381f33dd.

Bug: T352546
Change-Id: Ib9316b26035ed699f1e607f96e04da110c0edb32