Replace complicated strlen() checks in Sanitizer with is_int()
This is called quite often. But it's exceptionally rare that one of
the strings is actually to long. If we do the hexdec() conversion
first we can use a much cheaper is_int() check instead.

Note that codepoints larger than 0x10ffff are invalid anyway via the
validateCodepoint() check that's done in both cases. That's why this
patch is guaranteed to not make any difference, not even with the
two possible int sizes we can have in PHP.

Same as I7bbe2ea in the Parsoid codebase.

Change-Id: I4c4110b2f06f253bfeab988c593924d37c2e8c5d