Simplify Sanitizer::CHAR_REFS_REGEX regex a little
This regex is private and used in three places. They all already
use $matches[0] instead of $matches[4] for the last case (this is
different in the Parsoid codebase, see Ic25da24). The extra brackets
are not needed.

Also switch to checks for `null` instead of empty strings.

Same as Ic25da24 in the Parsoid codebase.

Change-Id: I038207b4fe665ad6c23911ec671594a9141aa232