Fix unanchored regex in jquery.highlightText.js
This was searching for a substring, which doesn't make sense in this
context.

This line of code is typically not doing anything because <style>
and <script> tags typically don't have children anyway. But this is
technically not forbidden. It's still a good idea to have this check
just to be sure.

This bug exists at least since 2010.

Change-Id: I4a18514168b9f3668487e2c95d477cad377b0be7