skins: do not fetch new talks when user has no active session
When rendering a page for anonymous user we shouldn't check for new talks
for user IP. Such request could be a CDN miss from someone who has an IP
of a contributor. When this happens and there are new talks for such IP
system would return a notification and make the page not cacheable.

Because we will show new talks only to users with a persistent session
lets remove disabling the CDN cache - as those request will always skip
the CDN cache and hit php servers.

Bug: T350861
Change-Id: Iec329617f9206c1cc0e9a2ae6bc04710f1273e75