tests: Skip expensive resetServices() when nothing changed
The `resetServices` call here is not the primary, documented effect
of `setMwGlobals` (as well as all the other functions that call
`setMwGlobals`). It's a necessary side-effect. But it's only
necessary when something actually changed.

A lot of tests do a lot of `setMwGlobals` calls just to be sure the
configuration is as expected. But more often than not it is already
fine.

Tests that depend on this `resetServices` call even when no
configuration change is made shouldn't rely on a hidden side-effect
but must do something else about this.

According to my local tests this makes a huge difference. I have
seen the runtime of test suites drop by as much as 1/3.

Change-Id: Id53c8da65d08730aa9536ea2ae71211c5170134a