mw.util: introduce isInfinity to parity MW's wfIsInfinity()
In our JS, especially around watchlisting, we use both 'infinity' and
'infinite' as infinity values, depending on if we're talking about the
expiry dropdown (which uses 'infinite') or API interaction (which more
commonly uses 'infinity').

This patch adds a stable method to test if a string is infinity so that
we don't need to worry about which context the value is coming from.

Bug: T353389
Change-Id: If53a88b20deb09998281794028704a64d1379f92