Document types in mw.notification.defaults
With the migration to JSDoc, the documentation for
`mw.notification.defaults` was no longer visible. Make it visible again
by using `@type` instead of `@property` and document its fields in a
machine-readable form. Put this machine-readable documentation in a
`@typedef` so that it can be referred to from the documentation of
`mw.notification.notify` and `Notification`.

Bug: T352308
Change-Id: If9762d040289cdc57d9153dc1fccba8047986713