CodexModule: Throw helpful error if component not listed in module definition
This patch adds a runtime check to identify when a Codex component is imported without first being defined in the `codexComponents` array of the module definition file.

The currently shown warnings or errors are not consistent between components and are confusing since they occur at different points in the code paths of each component.

The proposed change throws a custom error message mentioning the component missing from the module definition. The location of the component's `require` call can also be inferred from the stack trace.

Bug: T353187
Change-Id: Iee8e3b0769c10407ce08de35ba38e8e5a804d6c9