ParserOutput: Temporarily move "merge categories" in ::collectMetadata
Due to a botched signature change on the Parsoid side, in -a8 Parsoid
only accepts `string|int` for ContentMetadataCollector::addCategory()
and in -a9 Parsoid only accept `LinkTarget`.  The ParserOutput in
core, of course, accepts both.  So move the code which merges
categories into the section of ContentMetadataCollector::collectMetadata()
where we know that the CMC we're merging with is really a ParserOutput.

Change-Id: I0715f4fbc870e401e5759dd7c7a3c19077c40a6a