CodexModule: Handle empty codex component definitions
This patch handles the scenario of an empty codexComponents array in the
module definition file with a descriptive error message that allows
developers to recover from the situation.

The scenario of codexComponents being omitted entirely is not handled
since there is a way to use this module only for its themeStyles feature
that does not require definition of codexComponents.

Bug: T353187
Change-Id: I65ad8dc338f5e570e93d7918b053b953755c12d5