Skin: Restore autoloading of mediawiki.ui.button styles
This reverts commit 2a3ed74cbe9ee6d1ab59f3fea98546bb1abe995d.

The on-wiki use of these styles was the sole purpose of this Skin
autoloading of default styles based on content, similar to
"mw-collapsible" and "sortable".

This "deprecation" was in fact a breaking change that was unannounced,
unprepared, clearly unready, and without benefit to anyone.
It is already optimised to not load on articles.

Once a replaced is ready, we can slowly phase that in through the
same means, possibly using a stat counter to detect when it is ready
to remove. The impact of keeping is naturally low and mutually-exclusive
with breaking a page, as it will by design never load on pages where
it isn't needed.

If the CSS size of these non-content pages were a monitored performance
metric (not the case, as of writing, as these are workflow pages, not
articles), there are plenty trivial other optimisations that could be
made without needless disruption and far bigger net-impact.

Bug: T182050
Bug: T346469
Change-Id: Ice6d6cfb442e4b3edb7f3002799780c76d7e0b0e