Don't use Parsoid\Config\PageConfig::getTitle()
This has been replaced by ::getLinkTarget(), which returns a Parsoid
LinkTarget.  This is identical to the core LinkTarget interface, but
we can't quite alias them for technical reasons (sigh).  In actual
practice, LinkTargets generated by core are usually Title objects, so
Title::newFromLinkTarget() is a no-op that just returns the argument
after a type check.

It appears that newer code uses a TitleFormatter rather than calling
methods on Title, but TitleFormatter currently takes LinkTarget not a
ParsoidLinkTarget.  That would force us to go via
TitleValue::newFromLinkTarget() which isn't a simple type check.

Change-Id: I490bb38108d0202b43ea2a9b391b2e664e7d2d48