Add data-nosnippet to the sitenotice container
Substantial sitenotices were showing up in search previews, which is a
suboptimal preview for the article page.

Because of how the notices are built, this won't affect notices added by
extensions using the SiteNoticeAfter hook. Doing so would either require
an extra wrapper, or adding the attribute to every skin that outputs the
sitenotice.

Bug: T353984
Change-Id: I64f763c11e936ca781121fe59cbd919b84d37809