Fix classes when marking Special:Watchlist entries as seen
On the enhanced recent changes page, entries have the CSS classes
"mw-changeslist-line-watched" or "mw-changeslist-line-not-watched"
indicating whether the page is on the current user's watchlist.

On enhanced watchlist, the same CSS classes are re-used to indicate
whether the change has been seen by the user, rather than whether the
page is watched, as explained in a TODO comment here:
https://gerrit.wikimedia.org/g/mediawiki/core/+/20aace9da89c8661798ec9ba9882775d4f064d89/includes/specials/SpecialWatchlist.php#632

Either way, one of these two CSS classes should always be present,
since existing CSS relies on that (including on-wiki CSS).

The script that marks a single entry as seen after clicking it
did not ensure that, removing a class instead of swapping them.
Fixing this should fix the last CSS problems reported in T352456.

For comparison:
When generating the initial HTML, we always add one of the classes:
https://gerrit.wikimedia.org/g/mediawiki/core/+/20aace9da89c8661798ec9ba9882775d4f064d89/includes/changes/EnhancedChangesList.php#167
Watchlist script which marks *all* entries as seen swaps the classes:
https://gerrit.wikimedia.org/g/mediawiki/core/+/20aace9da89c8661798ec9ba9882775d4f064d89/resources/src/mediawiki.special.watchlist/watchlist.js#39

Bug: T352456
Change-Id: If3f303a9a71bd05c7cbaaf2f14924ce326da6a9d